image_to_text/requirements.txt
accelerate==1.3.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fairscale==0.4.13
filelock==3.17.0
fire==0.7.0
fsspec==2024.12.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
pip-tools==7.4.1
psutil==6.1.1
pyproject_hooks==1.2.0 
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
